Add credits UI
Lists all GitHub contributors and moderators, along with GitHub and TPT avatars. Avatars are embedded into the game for instant loading and to guarantee it looks right even without a network connection.

The UI itself is controlled with credits.json. This can be generated in advance using resources/gencredits.py, that script also downloads all the avatars.